Course runner. Now the information can be organised in courses/modules/lessons with multiple rhythm player controls at one lesson page.

+++ public/js/instrument-rack/instrument-rack-ui.js
@@ -114,18 +114,26 @@ class InstrumentRackUIController {
     }
 
     onClickXButton(id) {
+
+        // if player is playing - don't do anything
+        if ( mtRhythmPlayer.isPlaying ) return;
+
         this.rack.removeInstrumentInstance(id);
         this.render();
         this.rackChangedNotifier.notify( this.rack );
     }
 
     onClickDeleteAll() {
+
+        // if player is playing - don't add
+        if ( mtRhythmPlayer.isPlaying ) return;
+
         this.rack.deleteAll();
         this.render();
         this.rackChangedNotifier.notify( this.rack );
     }
 
-    generateInstrumentOptions() {
+    generateChoiceOfInstruments() {
         let select = document.getElementById( INSTRUMENT_SELECT_EL_ID );
         instrumentManager.allInstruments.forEach( instr => {
             let option = document.createElement("option");
